refactor: implement separate class for navigation work directory
