+++ src/index.js
@@ -1,4 +1,3 @@
 import { App } from "./app/App.js";
 
-
 new App();
